checked

+++ client/src/components/CommentSection.jsx
@@ -1,29 +1,46 @@
-import { Alert, Button, Modal, TextInput, Textarea } from 'flowbite-react';
-import { useEffect, useState } from 'react';
-import { useSelector } from 'react-redux';
-import { Link, useNavigate } from 'react-router-dom';
-import Comment from './comment'
-import { HiOutlineExclamationCircle } from 'react-icons/hi';
 
-export default function CommentSection({ postId }) {
+import { Alert, Button, Modal, Textarea } from "flowbite-react";
+import { useEffect, useState } from "react";
+import { useSelector } from "react-redux";
+import { Link, useNavigate } from "react-router-dom";
+import Comment from "./comment";
+import { HiOutlineExclamationCircle } from "react-icons/hi";
+
+const CommentSection = ({ postId }) => {
   const { currentUser } = useSelector((state) => state.user);
-  const [comment, setComment] = useState('');
+  const [comment, setComment] = useState("");
   const [commentError, setCommentError] = useState(null);
   const [comments, setComments] = useState([]);
   const [showModal, setShowModal] = useState(false);
   const [commentToDelete, setCommentToDelete] = useState(null);
-  const navigate = useNavigate(); 
+  const navigate = useNavigate();
+
+  useEffect(() => {
+    const getComments = async () => {
+      try {
+        const res = await fetch(`/api/comment/getPostComments/${postId}`);
+        const data = await res.json();
+
+        if (res.ok) {
+          setComments(data);
+        }
+      } catch (error) {
+        console.log(error.message);
+      }
+    };
+
+    getComments();
+  }, [postId]);
+
   const handleSubmit = async (e) => {
     e.preventDefault();
     if (comment.length > 200) {
       return;
     }
     try {
-      const res = await fetch('/api/comment/create', {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-        },
+      const res = await fetch("/api/comment/create", {
+        method: "POST",
+        headers: { "Content-Type": "application/json" },
         body: JSON.stringify({
           content: comment,
           postId,
@@ -31,39 +48,27 @@ export default function CommentSection({ postId }) {
         }),
       });
       const data = await res.json();
+
       if (res.ok) {
-        setComment('');
+        setComment("");
         setCommentError(null);
         setComments([data, ...comments]);
+      } else {
+        setCommentError(data.message);
       }
     } catch (error) {
       setCommentError(error.message);
     }
   };
 
-  useEffect(() => {
-    const getComments = async () => {
-      try {
-        const res = await fetch(`/api/comment/getPostComments/${postId}`);
-        if (res.ok) {
-          const data = await res.json();
-          setComments(data);
-        }
-      } catch (error) {
-        console.log(error.message);
-      }
-    };
-    getComments();
-  }, [postId]);
-
   const handleLike = async (commentId) => {
     try {
       if (!currentUser) {
-        navigate('/sign-in');
+        navigate("/sign-in");
         return;
       }
       const res = await fetch(`/api/comment/likeComment/${commentId}`, {
-        method: 'PUT',
+        method: "PUT",
       });
       if (res.ok) {
         const data = await res.json();
@@ -96,79 +101,83 @@ export default function CommentSection({ postId }) {
     setShowModal(false);
     try {
       if (!currentUser) {
-        navigate('/sign-in');
+        navigate("/sign-in");
         return;
       }
-      const res = await fetch(`/api/comment/deleteComment/${commentId}`, {
-        method: 'DELETE',
+      const res = await fetch(`/api/comment/delete/${commentId}`, {
+        method: "DELETE",
       });
+
       if (res.ok) {
         const data = await res.json();
+
         setComments(comments.filter((comment) => comment._id !== commentId));
       }
     } catch (error) {
       console.log(error.message);
     }
   };
+
   return (
-    <div className='max-w-2xl mx-auto w-full p-3'>
+    <div className="max-w-2xl mx-auto w-full p-3">
       {currentUser ? (
-        <div className='flex items-center gap-1 my-5 text-gray-500 text-sm'>
-          <p>Signed in as:</p>
+        <div className="flex items-center gap-1 my-5 text-gray-500 text-sm">
+          <p>Signed in as: </p>
           <img
-            className='h-5 w-5 object-cover rounded-full'
             src={currentUser.profilePicture}
-            alt=''
+            alt={currentUser.username}
+            className="h-5 w-5 object-cover rounded-full"
           />
           <Link
-            to={'/dashboard?tab=profile'}
-            className='text-xs text-cyan-600 hover:underline'
+            to={"/dashboard?tab=profile"}
+            className="text-xs text-cyan-600 hover:underline"
           >
             @{currentUser.username}
           </Link>
         </div>
       ) : (
-        <div className='text-sm text-teal-500 my-5 flex gap-1'>
-          You must be signed in to comment.
-          <Link className='text-blue-500 hover:underline' to={'/sign-in'}>
+        <div className="text-sm text-teal-500 my-5 flex gap-1">
+          You must be signed in to comment
+          <Link to={"/sign-in"} className="text-blue-500 hover:underline">
             Sign In
           </Link>
         </div>
       )}
+
       {currentUser && (
         <form
           onSubmit={handleSubmit}
-          className='border border-teal-500 rounded-md p-3'
+          className="border border-teal-500 rounded-md p-3 "
         >
           <Textarea
-            placeholder='Add a comment...'
-            rows='3'
-            maxLength='200'
+            placeholder="Add a comment..."
+            rows="3"
+            maxLength="200"
             onChange={(e) => setComment(e.target.value)}
             value={comment}
           />
-          <div className='flex justify-between items-center mt-5'>
-            <p className='text-gray-500 text-xs'>
+          <div className="flex justify-between items-center mt-5">
+            <p className="text-gray-500 text-xs">
               {200 - comment.length} characters remaining
             </p>
-            <Button outline gradientDuoTone='purpleToBlue' type='submit'>
+            <Button outline gradientDueTone="purpleToBlue" type="submit">
               Submit
             </Button>
           </div>
           {commentError && (
-            <Alert color='failure' className='mt-5'>
+            <Alert color="failure" className="mt-5">
               {commentError}
             </Alert>
           )}
         </form>
       )}
       {comments.length === 0 ? (
-        <p className='text-sm my-5'>No comments yet!</p>
+        <p className="text-sm my-5">No comments yet!</p>
       ) : (
         <>
-          <div className='text-sm my-5 flex items-center gap-1'>
+          <div className="text-sm my-5 flex items-center gap-1">
             <p>Comments</p>
-            <div className='border border-gray-400 py-1 px-2 rounded-sm'>
+            <div className="border border-gray-400 py-1 px-2 rounded-sm">
               <p>{comments.length}</p>
             </div>
           </div>
@@ -186,33 +195,37 @@ export default function CommentSection({ postId }) {
           ))}
         </>
       )}
-      <Modal
-        show={showModal}
-        onClose={() => setShowModal(false)}
-        popup
-        size='md'
-      >
-        <Modal.Header />
-        <Modal.Body>
-          <div className='text-center'>
-            <HiOutlineExclamationCircle className='h-14 w-14 text-gray-400 dark:text-gray-200 mb-4 mx-auto' />
-            <h3 className='mb-5 text-lg text-gray-500 dark:text-gray-400'>
-              Are you sure you want to delete this comment?
-            </h3>
-            <div className='flex justify-center gap-4'>
-              <Button
-                color='failure'
-                onClick={() => handleDelete(commentToDelete)}
-              >
-                Yes, I'm sure
-              </Button>
-              <Button color='gray' onClick={() => setShowModal(false)}>
-                No, cancel
-              </Button>
+      {showModal && (
+        <Modal
+          show={showModal}
+          onClose={() => setShowModal(false)}
+          popup
+          size="md"
+        >
+          <Modal.Header />
+          <Modal.Body>
+            <div className="text-center">
+              <HiOutlineExclamationCircle className="h-14 w-14 text-gray-400 dark:text-gray-200 mb-4 mx-auto" />
+              <h3 className="mb-5 text-lg text-gray-500 dark:text-gray-400">
+                Are you sure you want to delete this comment?
+              </h3>
+              <div className="flex justify-center gap-4">
+                <Button
+                  color="failure"
+                  onClick={() => handleDelete(commentToDelete)}
+                >
+                  Yes,I&apos;m sure
+                </Button>
+                <Button color="gray" onClick={() => setShowModal(false)}>
+                  No, cancel
+                </Button>
+              </div>
             </div>
-          </div>
-        </Modal.Body>
-      </Modal>
+          </Modal.Body>
+        </Modal>
+      )}
     </div>
   );
-}
+};
+
+export default CommentSection;
